src/pythonLambdas/requirements.txt
boto3==1.26.0
pdfminer.six==20231228
